"Trim-down" of AddSeat/AddProduct validation files...
This only contains changes to the most recent AddSeat/AddProduct files as they were the only ones with an approved trimmed-down validation set.

Preserving the old files with a date refernce in the name so that in case any of that info or structure is required, it can be easily be copied back into the trim-down files.

Also eliminated some commented chunks from a few Java files, but no code functionality affected...